+++ squads/__init__.py
@@ -2,7 +2,7 @@
 
 from squads.app_settings import SQUADS_APP_NAME
 
-__version__ = "0.2.2"
+__version__ = "0.2.3"
 __title__ = "Squads"
 
 USER_AGENT_TEXT = f"{SQUADS_APP_NAME} v{__version__}"