+++ squads/api/schema.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from typing import Optional
 
 from ninja import Schema
 
@@ -16,9 +15,9 @@ class Pending(Schema):
     group_id: int
     user: str
     approved: bool
-    application: str
+    comment: str
     application_id: str
-    skills: Optional[str] = None
+    req_filters: bool
     created_at: datetime
 
 
@@ -26,7 +25,7 @@ class Members(Schema):
     group: str
     group_id: int
     user: str
-    has_required_skills: bool
+    req_filters: bool
     is_active: bool
     joined_at: datetime
     application_id: str