+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## \[0.2.3\] - 2024-06-11
+
+### Added
+
+- Squads Filter Check Tasks
+- Pending Field `req_filters`
+- New Tests
+- Log Info on State Changes, Deletion
+
+### Changed
+
+- Groups Renamed Fields
+- Memberships Renamed Fields
+- Pending Renamed Fields
+
 ## \[0.2.2\] - 2024-06-11
 
 ### Added